fix: custom error handlingfor different operations

+++ CRUD.py
@@ -1,5 +1,7 @@
+from pyexpat.errors import messages
+
 from flask import Flask , request
-from flask_restful import Resource, Api , marshal_with , fields
+from flask_restful import Resource, Api , marshal_with , fields , abort
 from flask_sqlalchemy import SQLAlchemy
 
 app= Flask(__name__)
@@ -50,12 +52,18 @@ class Item(Resource):
     @marshal_with(taskFields)
     def get(self,pk):
         task = Task.query.filter_by(id=pk).first()
+        if not task:
+            abort(404,message=f'Task with id {pk} does not exist')
         return task
 
     @marshal_with(taskFields)
     def put(self,pk):
-       data=request.json
        task = Task.query.filter_by(id=pk).first()
+       if not task:
+           abort(404,message=f'item with id {pk} does not exist')
+       data = request.json
+       if not data or 'name' not in data:
+           abort(404,message='name field missing')
        task.name=data['name']
        db.session.commit()
        return task
@@ -63,6 +71,8 @@ class Item(Resource):
     @marshal_with(taskFields)
     def delete(self, pk):
         task = Task.query.filter_by(id=pk).first()
+        if not task :
+            abort(404,message=f'item with id {pk} does not exist')
         db.session.delete(task)
         db.session.commit()
         tasks = Task.query.all()
